+++ help.inc.php
@@ -1,2 +1,0 @@
-<?php echo $I18N->msg('ckeditor_addon_help'); ?>
-<br />