+++ module/input.php
@@ -1,0 +1,1 @@
+<textarea class="ckeditor" data-ckeditor-profile="default" name="REX_INPUT_VALUE[1]">REX_VALUE[1]</textarea>