+++ module/lite/output.php
@@ -1,1 +1,0 @@
-REX_HTML_VALUE[1]