+++ pages/help/update.inc.php
@@ -1,4 +1,0 @@
-<?php
-
-echo rex_ckeditor_utils::getHtmlFromMDFile('UPDATE.md');
-