+++ uninstall.inc.php
@@ -1,3 +1,0 @@
-<?php
-$REX['ADDON']['install']['ckeditor'] = false;
-