+++ update.php
@@ -1,0 +1,10 @@
+<?php
+
+// delete all assets
+rex_dir::deleteFiles($this->getAssetsPath(), true);
+
+// copy assets
+rex_dir::copy($this->getPath('assets'), $this->getAssetsPath());
+
+// apply patches
+rex_ckeditor::applyPatches();